Move fast marching native source into trajectory package
