+++ .github/workflows/fast-marching.yml
@@ -11,7 +11,7 @@ on:
       - "pyproject.toml"
       - "setup.py"
       - "MANIFEST.in"
-      - "recovar/_fast_marching_native.cpp"
+      - "recovar/trajectory/_fast_marching_native.cpp"
       - "recovar/heterogeneity/fast_marching.py"
       - "tests/unit/test_fast_marching.py"
       - "tests/unit/test_fast_marching_native.py"
@@ -23,7 +23,7 @@ on:
       - "pyproject.toml"
       - "setup.py"
       - "MANIFEST.in"
-      - "recovar/_fast_marching_native.cpp"
+      - "recovar/trajectory/_fast_marching_native.cpp"
       - "recovar/heterogeneity/fast_marching.py"
       - "tests/unit/test_fast_marching.py"
       - "tests/unit/test_fast_marching_native.py"
