+++ .github/workflows/wheels.yml
@@ -10,7 +10,7 @@ on:
       - "pyproject.toml"
       - "setup.py"
       - "MANIFEST.in"
-      - "recovar/_fast_marching_native.cpp"
+      - "recovar/trajectory/_fast_marching_native.cpp"
       - "recovar/heterogeneity/fast_marching.py"
       - "tests/unit/test_fast_marching.py"
       - "tests/unit/test_fast_marching_native.py"
@@ -20,7 +20,7 @@ on:
       - "pyproject.toml"
       - "setup.py"
       - "MANIFEST.in"
-      - "recovar/_fast_marching_native.cpp"
+      - "recovar/trajectory/_fast_marching_native.cpp"
       - "recovar/heterogeneity/fast_marching.py"
       - "tests/unit/test_fast_marching.py"
       - "tests/unit/test_fast_marching_native.py"
