+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include recovar/_fast_marching_native.cpp
+include recovar/trajectory/_fast_marching_native.cpp
