+++ recovar/trajectory/__init__.py
@@ -1,0 +1,1 @@
+"""Trajectory-related helpers and native extensions."""
